Bug Fixes and added a file saving and loading system
Redid the main logic of the program made it easier on the user

Added a file loading and saving system (unfinished)

Changed the way the user chooses the extension, the user now uses a Number to select what kind of file the want to move Video/Audio etc there is no longer an option to manually enter the extension

+++ ConsoleUI/FileManager.cs
@@ -4,6 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.IO;
+using System.Reflection;
+using System.Linq.Expressions;
+using System.Text.RegularExpressions;
 
 namespace ConsoleUI
 {
@@ -30,23 +33,81 @@ namespace ConsoleUI
         {
             return Directory.GetFiles(rootPath, $"*.{Extension}");
         }
-        public static void MoveFiles(string newLocation, string Extension,string rootPath,string Destination)
+        public static void MoveFile(string newLocation, string Extension,string rootPath,string Destination)
         {
             newLocation = Path.Combine(Destination, newLocation);
             string[] filesToMove = GetFilesToMove(rootPath, Extension);
 
-            foreach (var file in filesToMove)
-            {
-                string newFileName = UnwantedTextRemover(Path.GetFileName(file), "y2mate.com - ");
-                newFileName = Path.Combine(newLocation, newFileName);
-                File.Move(file, newFileName);
-                Console.WriteLine("Moving File " + Path.GetFileName(file));
-            }
+                foreach (var file in filesToMove)
+                {
+                    string newFileName = UnwantedTextRemover(Path.GetFileName(file), "y2mate.com - ");
+                    newFileName = Path.Combine(newLocation, newFileName);
+                if (!File.Exists(newFileName))
+                {
+                    File.Move(file, newFileName);
+                    Console.WriteLine("Moving File " + Path.GetFileName(file));
+                }
+                else
+                {Console.WriteLine($"File :{ Path.GetFileName(file)} all ready exists skipping file " );}
+                }
         }
 
         private static string UnwantedTextRemover(string text, string unwanted)
         {
             return text.Contains(unwanted) ? text.Replace(unwanted, "") : text;
+        }
+        public static void saveData(List<string> fileDestinations)
+        {
+            string path = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location),"Data");
+            string fileDestinationPath = path + @"\fileDestinations.txt";
+            StringBuilder sb = new StringBuilder();
+            StreamWriter sw;
+
+            try
+            {
+                sw = new StreamWriter(fileDestinationPath, true, Encoding.UTF8);
+                foreach (var destination in fileDestinations)
+                {
+                    sb.Append(destination).Append(',');
+                }
+                sw.WriteLine(sb);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+            }
+            finally
+            {
+                sb.Clear();
+            }
+        }
+
+        public static List<string> loadData(string type)
+        {
+            List<string> dataList = new List<string>();
+            try
+            {
+                string path = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "Data");
+                string text = File.ReadAllText(path + type);
+
+                text = text.Replace(Environment.NewLine, string.Empty);
+                string[] dataArray = text.Split(',');
+
+                foreach (var data in dataArray)
+                {
+                    string dataUnescaped = Regex.Unescape(data);
+                    if (!string.IsNullOrEmpty(dataUnescaped))
+                    {
+                        dataList.Add(data);
+                    }
+                }
+                return dataList;
+            }
+            catch(Exception e)
+            {
+                Console.WriteLine(e);
+                return dataList;
+            }
         }
     }
 }
